blazor-purity-vanilla-functions/requirements.txt
azure-functions
mcstatus==4.0.0
